Fixed another round of instancing bugs.
The basic problem is that the remove+add paradigm for updates doesn't work,
since the add of the (updated) slot isn't guaranteed to work. Add+remove
doesn't work since that clobbers the registration structures.

I don't like this change, as it seems to make the algorithms much more
fragile, hostile to modifications and unreadable, but I really don't know
what the alternative would be.