Sleep normally when FPS-limited instead of waiting on the dispatch queue.
Motivating example: Triggering many events per frame would cause the loop
to compensate later with a long sleep due to the FPS-limiter mechanism.
While that could probably be "fixed", this is probably thw wanted behavior
anyway. The only real argument against it seems to be as a failsafe against
unintentionally setting an ususably low FPS limit, which *may* be something
to evaluate.