Remove the extra MainFrame main-thread.
I don't understand why I felt this was necessary to begin with when it
was introduced in 62a3ad978, and it seems to work just fine without
it. If anything, this should make the main-thread more likely to reach
System.exit in a way that doesn't exit prematuerly (before sessions
are closed &c&c).