Fix HashDirCache LOCK_VIOLATION bug on Windows.
Apparently, Windows file locks are mandatory, so holding the file prevents
read/write operations. Therefore, lock only on the first byte of the
file and ensure the header is always completely read before unlocking the
file and returning it for asynchronous use.

If anything, the final replacement operation on stores might not work
entirely properly, but that's an exercise for another day for now.