Remove Composer, as it has been largely replaced with Widget.pos and similar mechanisms.
Widget.pos isn't exactly without problems either, but I have found it
generally more agreeable than changing the dynamic parameters of
Composer back and forth all the time to achieve intended margins,
paddings, and the like.

It's still worth pursuing a better solution still, and it may or may
not take a form similar to Composer, but it would certainly be
different enough that this wouldn't be a useful base for it.