Rework GLObject usage to not require finalization.
This is a change in model in that it is only render-system ro references
that are implicitly cleaned by being GC'd, and the GL backend itself is
now expected to always reliably dispose of internally used GLObjects. This
should generally be the case anyway, and so shouldn't be a problem, but it
is a somewhat significant change nonetheless.