Allow GL objects to be re-used as long as they aren't actually deleted.
While a rendering object may have been disposed from the client side,
it may stil be reused in the rendertree by coming from slots where it
still remains, and some of such cases are difficult or unreasonable to
eliminate, like states being concurrently updated for other reasons.

Technically, it might be nice to be able to differentiate actual "new
client-side" uses of a disposed state from reuse, but that doesn't
mesh well with the architecture, and is probably not worth it.