Skip sorting bone-weights by bone-name and just always sort by weight instead.
The conditional sorting by weight was obviously intended as an optimization,
but given that, according to 31c72de5, bone weights need to be consistently
sorted anyway, and that the alternative is always sorting by name, this
should now be an optimization.

The idea with sorting by names was that sorting by weights might yield an
inconsistent order when two or more weights are exactly equal, but bones
are sorted by name in the resource format to begin with, and Arrays.sort is
guaranteed to be stable, so that should not be an issue after all.