Add a backup JOGLWrap for non-GL3 environments.
This is mainly to ensure that capabilities can at least be queried, so that
the environment can be examined and a more proper error can be generated.

Kind of ugly, it's not like I'd really like to have multiple GL
implementations just for the sake of it, but I'm not sure what would be
better.