Remove all calls to the access-controller and related interfaces.
I find this somewhat melancholic, since I always found Java's sandbox
mode quite interesting, but it was deprecated in Java 17, and for
reasons I cannot (unfortunately) fully disagree with, so it is what
it is.

Haven also hasn't actually run in a sandbox since JNLP was also
deprecated (see 0eff1328), so no current functionality should be
compromised, even on Java 8.