Fix for minimap zoom on highres mousewheel
getWheelRotation() produces mousewheel messages with 0 'clicks' on high
resolution mouse wheels, like MacBook touchpad. In this case minimap
has an incorrect behaviour, as this partial clicks will make minimap
view zoom-in.
As `zoomlevel` and `scale` are both integers, this is the minimal fix to
the described issue.
Alternative solutions which can be considered:
* Do not fire mousewheel event when number of clicks == 0
* Provide an output from `getPreciseWheelRotation()` as an additional
argument to `mousewheel` calls to allow widget to decide what to do with
high precision wheels. Additionally provide a fractional zoom levels for
minimap to allow a more precise zoom behavior for such mouse wheels